@@ -63,8 +63,8 @@
     "axios": "^0.18.0",
     "babel-polyfill": "^6.26.0",
     "jquery": "^3.5.0",
-    "react": "16.4.1",
-    "react-dom": "16.4.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-router": "3.2.0"
   }
 }
